🔥 refactor: remove all legacy backfill and migration code
No historical data exists — missing data is a bug, not something to
silently paper over with runtime backfills.
