+++ internal/scheduler/persistence.go
@@ -6,8 +6,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"log/slog"
-	"os"
-	"path/filepath"
+	"maps"
 	"time"
 
 	"github.com/CherryHQ/stella/pkg/db/sqlc"
@@ -58,95 +57,6 @@ func (s *Service) recordJobRun(ctx context.Context, id, orgID string, ranAt time
 // deleteJob removes a job from the database.
 func (s *Service) deleteJob(ctx context.Context, id, orgID string) error {
 	return s.q.DeleteSchedulerJob(ctx, sqlc.DeleteSchedulerJobParams{ID: id, OrgID: orgID})
-}
-
-// migrateJobsFile imports jobs from the legacy jobs.json file into the database
-// and removes the file. This is a one-time migration on first startup.
-func (s *Service) migrateJobsFile(ctx context.Context, dataPath string) error {
-	if dataPath == "" {
-		return nil
-	}
-	file := filepath.Join(dataPath, "jobs.json")
-	data, err := os.ReadFile(file)
-	if os.IsNotExist(err) {
-		return nil
-	}
-	if err != nil {
-		return fmt.Errorf("read legacy jobs.json: %w", err)
-	}
-
-	var jobs []Job
-	if err := json.Unmarshal(data, &jobs); err != nil {
-		return fmt.Errorf("parse legacy jobs.json: %w", err)
-	}
-	if len(jobs) == 0 {
-		_ = os.Remove(file)
-		return nil
-	}
-
-	// Legacy jobs may lack OrgID. Look up the first existing org before
-	// starting the transaction to avoid SQLite lock contention.
-	var fallbackOrgID string
-	_ = s.db.QueryRowContext(ctx, `SELECT id FROM auth_organization ORDER BY created_at ASC LIMIT 1`).Scan(&fallbackOrgID)
-
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("begin migration tx: %w", err)
-	}
-	defer func() { _ = tx.Rollback() }()
-
-	qtx := s.q.WithTx(tx)
-	for _, job := range jobs {
-		if job.OrgID == "" {
-			if fallbackOrgID == "" {
-				s.log.Warn("skipping legacy job with no org_id (no orgs exist)", "job_id", job.ID)
-				continue
-			}
-			job.OrgID = fallbackOrgID
-		}
-		if _, err := qtx.CreateSchedulerJob(ctx, createSchedulerJobParams(job)); err != nil {
-			return fmt.Errorf("migrate job %s: %w", job.ID, err)
-		}
-	}
-
-	if err := tx.Commit(); err != nil {
-		return fmt.Errorf("commit migration: %w", err)
-	}
-
-	_ = os.Remove(file)
-	s.log.Info("migrated legacy jobs.json to database", "count", len(jobs))
-	return nil
-}
-
-// migrateLegacyPluginJobs converts plugin-owned jobs that still use the old
-// reserved message envelope into first-class ownership columns.
-func (s *Service) migrateLegacyPluginJobs(ctx context.Context) error {
-	rows, err := s.q.ListAllSchedulerJobs(ctx)
-	if err != nil {
-		return fmt.Errorf("list scheduler jobs: %w", err)
-	}
-	for _, row := range rows {
-		if row.OwnerKind == JobOwnerPlugin {
-			continue
-		}
-		job := dbRowToJob(row)
-		pluginID, key, runtimeName, description, payload, ok := DecodePluginJob(job)
-		if !ok {
-			continue
-		}
-		job.OwnerKind = JobOwnerPlugin
-		job.PluginID = pluginID
-		job.JobKey = key
-		job.RuntimeName = runtimeName
-		job.Description = description
-		job.Payload = payload
-		job.Message = ""
-		job.UpdatedAt = time.Now().UTC()
-		if err := s.updateJob(ctx, job); err != nil {
-			return fmt.Errorf("migrate legacy plugin job %s: %w", job.ID, err)
-		}
-	}
-	return nil
 }
 
 func createSchedulerJobParams(job Job) sqlc.CreateSchedulerJobParams {
@@ -310,6 +220,15 @@ func decodePayload(raw string) map[string]any {
 		return map[string]any{}
 	}
 	return payload
+}
+
+func clonePayload(src map[string]any) map[string]any {
+	if len(src) == 0 {
+		return map[string]any{}
+	}
+	out := make(map[string]any, len(src))
+	maps.Copy(out, src)
+	return out
 }
 
 func nullableTime(t *time.Time) sql.NullString {
