+++ internal/server/tasks.go
@@ -363,10 +363,6 @@ func toAPITaskEvent(e sqlc.AgentTaskEvent) apiserver.AgentTaskEvent {
 		var detail map[string]any
 		if err := json.Unmarshal([]byte(e.Detail), &detail); err == nil && len(detail) > 0 {
 			ae.Detail = &detail
-		} else if err != nil {
-			// Legacy plain-string detail: wrap for API compatibility.
-			wrapped := map[string]any{"message": e.Detail}
-			ae.Detail = &wrapped
 		}
 	}
 	return ae
