+++ pkg/db/sqlc/settings_channel_agent.sql.go
@@ -60,30 +60,6 @@ func (q *Queries) GetChatAgent(ctx context.Context, arg GetChatAgentParams) (Set
 	return i, err
 }
 
-const getLegacyChatAgent = `-- name: GetLegacyChatAgent :one
-SELECT channel_id, platform, chat_id, agent_id, org_id, updated_at FROM settings_channel_agent WHERE channel_id = '' AND platform = ? AND chat_id = ? AND org_id = ?
-`
-
-type GetLegacyChatAgentParams struct {
-	Platform string `json:"platform"`
-	ChatID   string `json:"chat_id"`
-	OrgID    string `json:"org_id"`
-}
-
-func (q *Queries) GetLegacyChatAgent(ctx context.Context, arg GetLegacyChatAgentParams) (SettingsChannelAgent, error) {
-	row := q.db.QueryRowContext(ctx, getLegacyChatAgent, arg.Platform, arg.ChatID, arg.OrgID)
-	var i SettingsChannelAgent
-	err := row.Scan(
-		&i.ChannelID,
-		&i.Platform,
-		&i.ChatID,
-		&i.AgentID,
-		&i.OrgID,
-		&i.UpdatedAt,
-	)
-	return i, err
-}
-
 const listChatAgents = `-- name: ListChatAgents :many
 SELECT channel_id, platform, chat_id, agent_id, org_id, updated_at FROM settings_channel_agent WHERE org_id = ? ORDER BY channel_id, platform, chat_id
 `
